refactor: Flatten execute calls for more performance
