+++ src/instructions/rvc.rs
@@ -1,10 +1,7 @@
-use super::super::machine::Machine;
-use super::super::{Error, SP};
+use super::super::SP;
 use super::register::Register;
-use super::utils::{rd, update_register, x, xs};
-use super::{
-    blank_instruction, common, extract_opcode, Instruction, Itype, Rtype, Stype, Utype, MODULE_RVC,
-};
+use super::utils::{rd, x, xs};
+use super::{blank_instruction, Instruction, Itype, Rtype, Stype, Utype};
 use crate::instructions as insts;
 
 // Notice the location of rs2 in RVC encoding is different from full encoding
@@ -90,202 +87,6 @@ fn b_immediate(instruction_bits: u32) -> i32 {
         | xs(instruction_bits, 12, 1, 8)) as i32
 }
 
-pub fn execute<Mac: Machine>(inst: Instruction, machine: &mut Mac) -> Result<(), Error> {
-    let op = extract_opcode(inst);
-    let next_pc: Option<Mac::REG> = match op {
-        insts::OP_RVC_SUB => {
-            let i = Rtype(inst);
-            common::sub(machine, i.rd(), i.rs1(), i.rs2());
-            None
-        }
-        insts::OP_RVC_ADD => {
-            let i = Rtype(inst);
-            common::add(machine, i.rd(), i.rs1(), i.rs2());
-            None
-        }
-        insts::OP_RVC_XOR => {
-            let i = Rtype(inst);
-            common::xor(machine, i.rd(), i.rs1(), i.rs2());
-            None
-        }
-        insts::OP_RVC_OR => {
-            let i = Rtype(inst);
-            common::or(machine, i.rd(), i.rs1(), i.rs2());
-            None
-        }
-        insts::OP_RVC_AND => {
-            let i = Rtype(inst);
-            common::and(machine, i.rd(), i.rs1(), i.rs2());
-            None
-        }
-        // > C.SUBW (RV64/128; RV32 RES)
-        insts::OP_RVC_SUBW => {
-            let i = Rtype(inst);
-            common::subw(machine, i.rd(), i.rs1(), i.rs2());
-            None
-        }
-        // > C.ADDW (RV64/128; RV32 RES)
-        insts::OP_RVC_ADDW => {
-            let i = Rtype(inst);
-            common::addw(machine, i.rd(), i.rs1(), i.rs2());
-            None
-        }
-        insts::OP_RVC_ADDI => {
-            let i = Itype(inst);
-            common::addi(machine, i.rd(), i.rs1(), i.immediate_s());
-            None
-        }
-        insts::OP_RVC_ANDI => {
-            let i = Itype(inst);
-            common::andi(machine, i.rd(), i.rs1(), i.immediate_s());
-            None
-        }
-        insts::OP_RVC_ADDIW => {
-            let i = Itype(inst);
-            common::addiw(machine, i.rd(), i.rs1(), i.immediate_s());
-            None
-        }
-        insts::OP_RVC_SLLI => {
-            let i = Itype(inst);
-            common::slli(machine, i.rd(), i.rs1(), i.immediate());
-            None
-        }
-        insts::OP_RVC_SRLI => {
-            let i = Itype(inst);
-            common::srli(machine, i.rd(), i.rs1(), i.immediate());
-            None
-        }
-        insts::OP_RVC_SRAI => {
-            let i = Itype(inst);
-            common::srai(machine, i.rd(), i.rs1(), i.immediate());
-            None
-        }
-        insts::OP_RVC_LW => {
-            let i = Itype(inst);
-            common::lw(machine, i.rd(), i.rs1(), i.immediate_s())?;
-            None
-        }
-        insts::OP_RVC_LD => {
-            let i = Itype(inst);
-            common::ld(machine, i.rd(), i.rs1(), i.immediate_s())?;
-            None
-        }
-        insts::OP_RVC_SW => {
-            let i = Stype(inst);
-            common::sw(machine, i.rs1(), i.rs2(), i.immediate_s())?;
-            None
-        }
-        insts::OP_RVC_SD => {
-            let i = Stype(inst);
-            common::sd(machine, i.rs1(), i.rs2(), i.immediate_s())?;
-            None
-        }
-        insts::OP_RVC_LI => {
-            let i = Utype(inst);
-            update_register(machine, i.rd(), Mac::REG::from_i32(i.immediate_s()));
-            None
-        }
-        insts::OP_RVC_LUI => {
-            let i = Utype(inst);
-            update_register(machine, i.rd(), Mac::REG::from_i32(i.immediate_s()));
-            None
-        }
-        insts::OP_RVC_ADDI4SPN => {
-            let i = Utype(inst);
-            let value = machine.registers()[SP].overflowing_add(&Mac::REG::from_u32(i.immediate()));
-            update_register(machine, i.rd(), value);
-            None
-        }
-        insts::OP_RVC_LWSP => {
-            let i = Utype(inst);
-            common::lw(machine, i.rd(), SP, i.immediate_s())?;
-            None
-        }
-        insts::OP_RVC_LDSP => {
-            let i = Utype(inst);
-            common::ld(machine, i.rd(), SP, i.immediate_s())?;
-            None
-        }
-        insts::OP_RVC_SWSP => {
-            let i = Stype(inst);
-            common::sw(machine, SP, i.rs2(), i.immediate_s())?;
-            None
-        }
-        insts::OP_RVC_SDSP => {
-            let i = Stype(inst);
-            common::sd(machine, SP, i.rs2(), i.immediate_s())?;
-            None
-        }
-        insts::OP_RVC_BEQZ => {
-            let i = Stype(inst);
-            let condition = machine.registers()[i.rs1()].eq(&Mac::REG::zero());
-            let next_pc_offset = condition.cond(
-                &Mac::REG::from_i32(i.immediate_s()),
-                &Mac::REG::from_usize(2),
-            );
-            Some(machine.pc().overflowing_add(&next_pc_offset))
-        }
-        insts::OP_RVC_BNEZ => {
-            let i = Stype(inst);
-            let condition = machine.registers()[i.rs1()]
-                .eq(&Mac::REG::zero())
-                .logical_not();
-            let next_pc_offset = condition.cond(
-                &Mac::REG::from_i32(i.immediate_s()),
-                &Mac::REG::from_usize(2),
-            );
-            Some(machine.pc().overflowing_add(&next_pc_offset))
-        }
-        insts::OP_RVC_MV => {
-            let i = Rtype(inst);
-            let value = &machine.registers()[i.rs2()];
-            update_register(machine, i.rd(), value.clone());
-            None
-        }
-        insts::OP_RVC_JAL => {
-            let i = Utype(inst);
-            common::jal(machine, 1, i.immediate_s(), 2)
-        }
-        insts::OP_RVC_J => {
-            let i = Utype(inst);
-            Some(
-                machine
-                    .pc()
-                    .overflowing_add(&Mac::REG::from_i32(i.immediate_s())),
-            )
-        }
-        insts::OP_RVC_JR => {
-            let i = Stype(inst);
-            Some(machine.registers()[i.rs1()].clone())
-        }
-        insts::OP_RVC_JALR => {
-            let i = Stype(inst);
-            let link = machine.pc().overflowing_add(&Mac::REG::from_usize(2));
-            update_register(machine, 1, link);
-            Some(machine.registers()[i.rs1()].clone())
-        }
-        insts::OP_RVC_ADDI16SP => {
-            let i = Itype(inst);
-            let value =
-                machine.registers()[SP].overflowing_add(&Mac::REG::from_i32(i.immediate_s()));
-            update_register(machine, SP, value);
-            None
-        }
-        insts::OP_RVC_SRLI64 => None,
-        insts::OP_RVC_SRAI64 => None,
-        insts::OP_RVC_SLLI64 => None,
-        insts::OP_RVC_NOP => None,
-        insts::OP_RVC_EBREAK => {
-            machine.ebreak()?;
-            None
-        }
-        _ => return Err(Error::InvalidOp(op as u8)),
-    };
-    let default_next_pc = machine.pc().overflowing_add(&Mac::REG::from_usize(2));
-    machine.set_pc(next_pc.unwrap_or(default_next_pc));
-    Ok(())
-}
-
 #[allow(clippy::cyclomatic_complexity)]
 pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
     let bit_length = R::BITS;
@@ -307,7 +108,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                         insts::OP_RVC_ADDI4SPN,
                         compact_register_number(instruction_bits, 2),
                         nzuimm,
-                        MODULE_RVC,
                     )
                     .0,
                 )
@@ -322,7 +122,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                 compact_register_number(instruction_bits, 2),
                 compact_register_number(instruction_bits, 7),
                 sw_uimmediate(instruction_bits),
-                MODULE_RVC,
             )
             .0,
         ),
@@ -336,7 +135,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                         compact_register_number(instruction_bits, 2),
                         compact_register_number(instruction_bits, 7),
                         fld_uimmediate(instruction_bits),
-                        MODULE_RVC,
                     )
                     .0,
                 )
@@ -350,7 +148,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                 sw_uimmediate(instruction_bits),
                 compact_register_number(instruction_bits, 7),
                 compact_register_number(instruction_bits, 2),
-                MODULE_RVC,
             )
             .0,
         ),
@@ -364,7 +161,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                         fld_uimmediate(instruction_bits),
                         compact_register_number(instruction_bits, 7),
                         compact_register_number(instruction_bits, 2),
-                        MODULE_RVC,
                     )
                     .0,
                 )
@@ -375,9 +171,9 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
             let nzimm = immediate(instruction_bits);
             let rd = rd(instruction_bits);
             if nzimm != 0 && rd != 0 {
-                Some(Itype::new_s(insts::OP_RVC_ADDI, rd, rd, nzimm, MODULE_RVC).0)
+                Some(Itype::new_s(insts::OP_RVC_ADDI, rd, rd, nzimm).0)
             } else if nzimm == 0 && rd == 0 {
-                Some(blank_instruction(insts::OP_RVC_NOP, MODULE_RVC))
+                Some(blank_instruction(insts::OP_RVC_NOP))
             } else {
                 // Invalid instruction
                 None
@@ -385,28 +181,11 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
         }
         0b_001_00000000000_01 => {
             if rv32 {
-                Some(
-                    Utype::new_s(
-                        insts::OP_RVC_JAL,
-                        0,
-                        j_immediate(instruction_bits),
-                        MODULE_RVC,
-                    )
-                    .0,
-                )
+                Some(Utype::new_s(insts::OP_RVC_JAL, 0, j_immediate(instruction_bits)).0)
             } else {
                 let rd = rd(instruction_bits);
                 if rd != 0 {
-                    Some(
-                        Itype::new_s(
-                            insts::OP_RVC_ADDIW,
-                            rd,
-                            rd,
-                            immediate(instruction_bits),
-                            MODULE_RVC,
-                        )
-                        .0,
-                    )
+                    Some(Itype::new_s(insts::OP_RVC_ADDIW, rd, rd, immediate(instruction_bits)).0)
                 } else {
                     None
                 }
@@ -415,15 +194,7 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
         0b_010_00000000000_01 => {
             let rd = rd(instruction_bits);
             if rd != 0 {
-                Some(
-                    Utype::new_s(
-                        insts::OP_RVC_LI,
-                        rd,
-                        immediate(instruction_bits),
-                        MODULE_RVC,
-                    )
-                    .0,
-                )
+                Some(Utype::new_s(insts::OP_RVC_LI, rd, immediate(instruction_bits)).0)
             } else {
                 None
             }
@@ -444,12 +215,11 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                                 | x(instruction_bits, 3, 2, 7)
                                 | xs(instruction_bits, 12, 1, 9))
                                 as i32,
-                            MODULE_RVC,
                         )
                         .0,
                     )
                 } else if rd != 0 {
-                    Some(Utype::new_s(insts::OP_RVC_LUI, rd, imm, MODULE_RVC).0)
+                    Some(Utype::new_s(insts::OP_RVC_LUI, rd, imm).0)
                 } else {
                     None
                 }
@@ -462,11 +232,11 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
             match instruction_bits & 0b_1_11_000_11000_00 {
                 // SRLI64
                 0b_0_00_000_00000_00 if instruction_bits & 0b_111_00 == 0 => {
-                    Some(blank_instruction(insts::OP_RVC_SRLI64, MODULE_RVC))
+                    Some(blank_instruction(insts::OP_RVC_SRLI64))
                 }
                 // SRAI64
                 0b_0_01_000_00000_00 if instruction_bits & 0b_111_00 == 0 => {
-                    Some(blank_instruction(insts::OP_RVC_SRAI64, MODULE_RVC))
+                    Some(blank_instruction(insts::OP_RVC_SRAI64))
                 }
                 // SUB
                 0b_0_11_000_00000_00 => Some(
@@ -475,7 +245,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                         rd,
                         rd,
                         compact_register_number(instruction_bits, 2),
-                        MODULE_RVC,
                     )
                     .0,
                 ),
@@ -486,7 +255,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                         rd,
                         rd,
                         compact_register_number(instruction_bits, 2),
-                        MODULE_RVC,
                     )
                     .0,
                 ),
@@ -497,7 +265,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                         rd,
                         rd,
                         compact_register_number(instruction_bits, 2),
-                        MODULE_RVC,
                     )
                     .0,
                 ),
@@ -508,7 +275,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                         rd,
                         rd,
                         compact_register_number(instruction_bits, 2),
-                        MODULE_RVC,
                     )
                     .0,
                 ),
@@ -519,7 +285,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                         rd,
                         rd,
                         compact_register_number(instruction_bits, 2),
-                        MODULE_RVC,
                     )
                     .0,
                 ),
@@ -530,7 +295,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                         rd,
                         rd,
                         compact_register_number(instruction_bits, 2),
-                        MODULE_RVC,
                     )
                     .0,
                 ),
@@ -545,46 +309,32 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                         (0b_00_000_00000_00, 0) => None,
                         // SRLI
                         (0b_00_000_00000_00, uimm) => {
-                            Some(Itype::new(insts::OP_RVC_SRLI, rd, rd, uimm, MODULE_RVC).0)
+                            Some(Itype::new(insts::OP_RVC_SRLI, rd, rd, uimm).0)
                         }
                         // Invalid instruction
                         (0b_01_000_00000_00, 0) => None,
                         // SRAI
                         (0b_01_000_00000_00, uimm) => {
-                            Some(Itype::new(insts::OP_RVC_SRAI, rd, rd, uimm, MODULE_RVC).0)
+                            Some(Itype::new(insts::OP_RVC_SRAI, rd, rd, uimm).0)
                         }
                         // ANDI
                         (0b_10_000_00000_00, _) => Some(
-                            Itype::new_s(
-                                insts::OP_RVC_ANDI,
-                                rd,
-                                rd,
-                                immediate(instruction_bits),
-                                MODULE_RVC,
-                            )
-                            .0,
+                            Itype::new_s(insts::OP_RVC_ANDI, rd, rd, immediate(instruction_bits)).0,
                         ),
                         _ => None,
                     }
                 }
             }
         }
-        0b_101_00000000000_01 => Some(
-            Utype::new_s(
-                insts::OP_RVC_J,
-                0,
-                j_immediate(instruction_bits),
-                MODULE_RVC,
-            )
-            .0,
-        ),
+        0b_101_00000000000_01 => {
+            Some(Utype::new_s(insts::OP_RVC_J, 0, j_immediate(instruction_bits)).0)
+        }
         0b_110_00000000000_01 => Some(
             Stype::new_s(
                 insts::OP_RVC_BEQZ,
                 b_immediate(instruction_bits),
                 compact_register_number(instruction_bits, 7),
                 0,
-                MODULE_RVC,
             )
             .0,
         ),
@@ -594,7 +344,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                 b_immediate(instruction_bits),
                 compact_register_number(instruction_bits, 7),
                 0,
-                MODULE_RVC,
             )
             .0,
         ),
@@ -606,23 +355,15 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                 // Reserved
                 None
             } else if uimm != 0 {
-                Some(Itype::new(insts::OP_RVC_SLLI, rd, rd, uimm, MODULE_RVC).0)
+                Some(Itype::new(insts::OP_RVC_SLLI, rd, rd, uimm).0)
             } else {
-                Some(blank_instruction(insts::OP_RVC_SLLI64, MODULE_RVC))
+                Some(blank_instruction(insts::OP_RVC_SLLI64))
             }
         }
         0b_010_00000000000_10 => {
             let rd = rd(instruction_bits);
             if rd != 0 {
-                Some(
-                    Utype::new(
-                        insts::OP_RVC_LWSP,
-                        rd,
-                        lwsp_uimmediate(instruction_bits),
-                        MODULE_RVC,
-                    )
-                    .0,
-                )
+                Some(Utype::new(insts::OP_RVC_LWSP, rd, lwsp_uimmediate(instruction_bits)).0)
             } else {
                 // Reserved
                 None
@@ -634,15 +375,7 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
             } else {
                 let rd = rd(instruction_bits);
                 if rd != 0 {
-                    Some(
-                        Utype::new(
-                            insts::OP_RVC_LDSP,
-                            rd,
-                            fldsp_uimmediate(instruction_bits),
-                            MODULE_RVC,
-                        )
-                        .0,
-                    )
+                    Some(Utype::new(insts::OP_RVC_LDSP, rd, fldsp_uimmediate(instruction_bits)).0)
                 } else {
                     // Reserved
                     None
@@ -657,20 +390,18 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                     if rd == 0 {
                         None
                     } else if rs2 == 0 {
-                        Some(Stype::new(insts::OP_RVC_JR, 0, rd, 0, MODULE_RVC).0)
+                        Some(Stype::new(insts::OP_RVC_JR, 0, rd, 0).0)
                     } else {
-                        Some(Rtype::new(insts::OP_RVC_MV, rd, 0, rs2, MODULE_RVC).0)
+                        Some(Rtype::new(insts::OP_RVC_MV, rd, 0, rs2).0)
                     }
                 }
                 0b_1_00000_00000_00 => {
                     let rd = rd(instruction_bits);
                     let rs2 = c_rs2(instruction_bits);
                     match (rd, rs2) {
-                        (0, 0) => Some(blank_instruction(insts::OP_RVC_EBREAK, MODULE_RVC)),
-                        (rs1, 0) => Some(Stype::new(insts::OP_RVC_JALR, 0, rs1, 0, MODULE_RVC).0),
-                        (rd, rs2) if rd != 0 => {
-                            Some(Rtype::new(insts::OP_RVC_ADD, rd, rd, rs2, MODULE_RVC).0)
-                        }
+                        (0, 0) => Some(blank_instruction(insts::OP_RVC_EBREAK)),
+                        (rs1, 0) => Some(Stype::new(insts::OP_RVC_JALR, 0, rs1, 0).0),
+                        (rd, rs2) if rd != 0 => Some(Rtype::new(insts::OP_RVC_ADD, rd, rd, rs2).0),
                         // Invalid instruction
                         _ => None,
                     }
@@ -684,7 +415,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                 swsp_uimmediate(instruction_bits),
                 0,
                 c_rs2(instruction_bits),
-                MODULE_RVC,
             )
             .0,
         ),
@@ -698,7 +428,6 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<Instruction> {
                         fsdsp_uimmediate(instruction_bits),
                         0,
                         c_rs2(instruction_bits),
-                        MODULE_RVC,
                     )
                     .0,
                 )
